+++ .gitignore
@@ -1,3 +1,3 @@
 # Virtual environment
 venv/
-
+node_modules/
